# specs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Azure AI Services Solutions",
    "description": "This contains a collection of solutions that leverage Azure AI services.",
    "version": "0.0.1",
    "x-logo": {
      "url": "https://news.microsoft.com/wp-content/uploads/prod/2022/05/Microsoft-logo_rgb_c-gray-1024x459.png"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "paths": {
    "/azure_openai/chat_completions/": {
      "post": {
        "tags": [
          "azure_openai"
        ],
        "summary": "Create Chat Completions",
        "operationId": "create_chat_completions_azure_openai_chat_completions__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/document_intelligence/analyze_document/": {
      "post": {
        "tags": [
          "document_intelligence"
        ],
        "summary": "Analyze Document",
        "operationId": "analyze_document_document_intelligence_analyze_document__post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_document_document_intelligence_analyze_document__post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeDocumentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/azure_storage/blobs/upload/": {
      "post": {
        "tags": [
          "azure_storage"
        ],
        "summary": "Upload Blob",
        "operationId": "upload_blob_azure_storage_blobs_upload__post",
        "parameters": [
          {
            "name": "blob_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Blob Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_blob_azure_storage_blobs_upload__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlobUploadResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/azure_storage/blobs/delete/": {
      "delete": {
        "tags": [
          "azure_storage"
        ],
        "summary": "Delete Blob",
        "operationId": "delete_blob_azure_storage_blobs_delete__delete",
        "parameters": [
          {
            "name": "blob_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Blob Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/azure_storage/blobs/": {
      "get": {
        "tags": [
          "azure_storage"
        ],
        "summary": "List Blobs",
        "operationId": "list_blobs_azure_storage_blobs__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalyzeDocumentResponse": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "AnalyzeDocumentResponse"
      },
      "BlobUploadResponse": {
        "properties": {
          "blob_name": {
            "type": "string",
            "title": "Blob Name"
          }
        },
        "type": "object",
        "required": [
          "blob_name"
        ],
        "title": "BlobUploadResponse"
      },
      "Body_analyze_document_document_intelligence_analyze_document__post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_analyze_document_document_intelligence_analyze_document__post"
      },
      "Body_upload_blob_azure_storage_blobs_upload__post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_blob_azure_storage_blobs_upload__post"
      },
      "ChatCompletionRequest": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ChatCompletionRequest"
      },
      "ChatCompletionResponse": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ChatCompletionResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}